+++ presenter/src/main/java/com/mashup/twotoo/presenter/home/Home.kt
@@ -4,7 +4,6 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
@@ -31,7 +30,7 @@ fun HomeRoute(
 
 @Composable
 fun HomeScreen(
-    onBeeButtonClick : () -> Unit = {},
+    onBeeButtonClick: () -> Unit = {},
     modifier: Modifier = Modifier,
     challengeStateTypeUiModel: ChallengeStateTypeUiModel = BeforeChallengeUiModel.empty,
 ) {
