+++ presenter/src/main/java/com/mashup/twotoo/presenter/home/navigation/HomeNavigation.kt
@@ -1,5 +1,7 @@
 package com.mashup.twotoo.presenter.home.navigation
 
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.ui.Modifier
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
@@ -13,6 +15,8 @@ fun NavController.navigateToHome(navOptions: NavOptions? = null) {
 }
 fun NavGraphBuilder.homeGraph() {
     composable(route = HomeNavigationRoute) {
-        HomeRoute()
+        HomeRoute(
+            modifier = Modifier.fillMaxSize(),
+        )
     }
 }
