+++ presenter/src/main/java/com/mashup/twotoo/presenter/home/ongoing/components/HomeGoalAchievement.kt
@@ -1,4 +1,4 @@
-package com.mashup.twotoo.presenter.home.ongoing
+package com.mashup.twotoo.presenter.home.ongoing.components
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.LinearProgressIndicator
